feat(admin): extract API error handler hook
